requirements.txt
pandas
opensearch-py
python-dotenv